Added Async pipes for windows and minor fixes to the build
Added support for more command line options and Fixed song playing error

Fixed a spelling mistake in a comment in Pipe.cpp
It now can check all command line arguments and added an argument named -subdir so that a user can specify the a specific folder within the music folder.
Fixed an error with getCurrentSong() failing in the play_next loop because of the playedFiles vector being cleared.

Changed for loop for consistency

Fixed Windows Pipe

The error, after investigation, seemed to be because of a Pipe not being disconnected after there were errors.

Added support for more command line options and Fixed song playing error (#25)

* Added Async pipes for windows and minor fixes to the build

* Added support for more command line options and Fixed song playing error

Fixed a spelling mistake in a comment in Pipe.cpp
It now can check all command line arguments and added an argument named -subdir so that a user can specify the a specific folder within the music folder.
Fixed an error with getCurrentSong() failing in the play_next loop because of the playedFiles vector being cleared.

* Changed for loop for consistency

Added support for more command line options and Fixed song playing error (#16) (#26)

* Added Async pipes for windows and minor fixes to the build

* Added support for more command line options and Fixed song playing error

Fixed a spelling mistake in a comment in Pipe.cpp
It now can check all command line arguments and added an argument named -subdir so that a user can specify the a specific folder within the music folder.
Fixed an error with getCurrentSong() failing in the play_next loop because of the playedFiles vector being cleared.

* Changed for loop for consistency

Fix compile on Linux

Added -subdir command to readme.md